Move the project custom field calculation logic from the SetAttributesService to the Create, Update and Copy services
