+++ app/services/projects/concerns/new_project_service.rb
@@ -111,5 +111,23 @@ module Projects::Concerns
 
       project.project_custom_field_project_mappings.build(mappings)
     end
+
+    def update_calculated_value_custom_fields(model)
+      changed_cf_ids = model.custom_values.map(&:custom_field_id)
+
+      # Using unscope(where: :admin_only) to fix an issue when non admin user
+      # edits a custom field which is used by an admin only calculated value
+      # field. Without this unscoping, admin only value and all fields
+      # referencing it (recursively) will not be recalculated and there will
+      # even be no place for that recalculatin to be triggered unless an admin
+      # edits same value again.
+      #
+      # This may need to be handled differently to make it work for other custom
+      # field containers, like WorkPackage. User custom fields also has
+      # admin_only check.
+      affected_cfs = model.available_custom_fields.unscope(where: :admin_only).affected_calculated_fields(changed_cf_ids)
+
+      model.calculate_custom_fields(affected_cfs)
+    end
   end
 end
