+++ app/services/projects/copy_service.rb
@@ -100,6 +100,8 @@ module Projects
     def after_perform(call)
       super.tap do |super_call|
         copy_activated_custom_fields(super_call)
+        update_calculated_value_custom_fields(super_call.result)
+        super_call.result.save if super_call.result.changed_for_autosave?
       end
     end
