+++ app/services/projects/create_service.rb
@@ -31,6 +31,13 @@
 module Projects
   class CreateService < ::BaseServices::Create
     include Projects::Concerns::NewProjectService
-    include Projects::Concerns::SetCalculatedCustomFieldValues
+
+    def after_perform(service_call)
+      super.tap do |call|
+        new_project = call.result
+        update_calculated_value_custom_fields(new_project)
+        new_project.save if new_project.changed_for_autosave?
+      end
+    end
   end
 end
