+++ app/services/projects/update_service.rb
@@ -31,7 +31,6 @@
 module Projects
   class UpdateService < ::BaseServices::Update
     prepend Projects::Concerns::UpdateDemoData
-    include Projects::Concerns::SetCalculatedCustomFieldValues
 
     private
 
@@ -47,8 +46,15 @@ module Projects
       ret
     end
 
+    def before_perform(service_call)
+      super.tap do |super_call|
+        update_calculated_value_custom_fields(super_call.result)
+      end
+    end
+
     def after_perform(service_call)
       ret = super
+
       touch_on_custom_values_update
       notify_on_identifier_renamed
       send_update_notification
@@ -80,6 +86,24 @@ module Projects
       return unless memoized_changes["parent_id"]
 
       WorkPackage.update_versions_from_hierarchy_change(model)
+    end
+
+    def update_calculated_value_custom_fields(model)
+      changed_cf_ids = model.custom_values.select(&:changed?).map(&:custom_field_id)
+
+      # Using unscope(where: :admin_only) to fix an issue when non admin user
+      # edits a custom field which is used by an admin only calculated value
+      # field. Without this unscoping, admin only value and all fields
+      # referencing it (recursively) will not be recalculated and there will
+      # even be no place for that recalculatin to be triggered unless an admin
+      # edits same value again.
+      #
+      # This may need to be handled differently to make it work for other custom
+      # field containers, like WorkPackage. User custom fields also has
+      # admin_only check.
+      affected_cfs = model.available_custom_fields.unscope(where: :admin_only).affected_calculated_fields(changed_cf_ids)
+
+      model.calculate_custom_fields(affected_cfs)
     end
 
     def handle_archiving
